+++ packages/api/tests/endpoints/stats.js
@@ -2,41 +2,32 @@
 
 const test = require('ava');
 const request = require('supertest');
-const rewire = require('rewire');
-const sinon = require('sinon');
+const range = require('lodash/range');
 
 const awsServices = require('@cumulus/aws-client/services');
 const s3 = require('@cumulus/aws-client/S3');
 const { randomId } = require('@cumulus/common/test-utils');
-const { bootstrapElasticSearch } = require('@cumulus/es-client/bootstrap');
-const indexer = rewire('@cumulus/es-client/indexer');
-const { getEsClient } = require('@cumulus/es-client/search');
 
 const models = require('../../models');
 const {
-  fakeGranuleFactoryV2,
-  fakeCollectionFactory,
   createFakeJwtAuthToken,
   setAuthorizedOAuthUsers,
 } = require('../../lib/testUtils');
 
+const {
+  destroyLocalTestDb,
+  generateLocalTestDb,
+  GranulePgModel,
+  CollectionPgModel,
+  fakeCollectionRecordFactory,
+  fakeGranuleRecordFactory,
+  migrationDir,
+  localStackConnectionEnv,
+} = require('../../../db/dist');
+
+const testDbName = randomId('collection');
+
 const assertions = require('../../lib/assertions');
-
-const stats = rewire('../../endpoints/stats');
-const getType = stats.__get__('getType');
-
-let esClient;
-
-process.env.AccessTokensTable = randomId('accessTokenTable');
-
-process.env.system_bucket = randomId('bucket');
-process.env.stackName = randomId('stackName');
-
-const esIndex = randomId('esindex');
-const esAlias = randomId('esAlias');
-
-process.env.ES_INDEX = esAlias;
-process.env.TOKEN_SECRET = randomId('tokensecret');
 
 // import the express app after setting the env variables
 const { app } = require('../../app');
@@ -44,10 +35,24 @@ const { app } = require('../../app');
 let accessTokenModel;
 let jwtAuthToken;
 
-test.before(async () => {
+process.env.PG_HOST = randomId('hostname');
+process.env.PG_USER = randomId('user');
+process.env.PG_PASSWORD = randomId('password');
+process.env.stackName = randomId('userstack');
+process.env.AccessTokensTable = randomId('tokentable');
+process.env.system_bucket = randomId('bucket');
+process.env.stackName = randomId('stackName');
+process.env.TOKEN_SECRET = randomId('tokensecret');
+
+process.env = {
+  ...process.env,
+  ...localStackConnectionEnv,
+  PG_DATABASE: testDbName,
+};
+
+test.before(async (t) => {
   // create buckets
   await awsServices.s3().createBucket({ Bucket: process.env.system_bucket });
-  esClient = await getEsClient();
   const username = randomId();
   await setAuthorizedOAuthUsers([username]);
 
@@ -56,48 +61,63 @@ test.before(async () => {
 
   jwtAuthToken = await createFakeJwtAuthToken({ accessTokenModel, username });
 
-  // create the elasticsearch index and add mapping
-  await bootstrapElasticSearch({
-    host: 'fakehost',
-    index: esIndex,
-    alias: esAlias,
-  });
-  // Index test data - 2 collections, 3 granules
-  await Promise.all([
-    indexer.indexCollection(esClient, fakeCollectionFactory(), esAlias),
-    indexer.indexCollection(esClient, fakeCollectionFactory(), esAlias),
-    indexer.indexGranule(esClient, fakeGranuleFactoryV2({ collectionId: 'coll1' }), esAlias),
-    indexer.indexGranule(esClient, fakeGranuleFactoryV2({ collectionId: 'coll1' }), esAlias),
-    indexer.indexGranule(esClient, fakeGranuleFactoryV2({ status: 'failed', duration: 3 }), esAlias),
-  ]);
+  const { knexAdmin, knex } = await generateLocalTestDb(
+    testDbName,
+    migrationDir
+  );
 
-  // Indexing using Date.now() to generate the timestamp
-  const stub = sinon.stub(Date, 'now').returns((new Date(2020, 0, 29)).getTime());
+  t.context.knexAdmin = knexAdmin;
+  t.context.knex = knex;
 
-  await Promise.all([
-    indexer.indexCollection(esClient, fakeCollectionFactory({
-      updatedAt: new Date(2020, 0, 29),
-    }), esAlias),
-    indexer.indexGranule(esClient, fakeGranuleFactoryV2({
-      status: 'failed',
-      updatedAt: new Date(2020, 0, 29),
-      duration: 4,
-    }), esAlias),
-    indexer.indexGranule(esClient, fakeGranuleFactoryV2({
-      updatedAt: new Date(2020, 0, 29),
-      duration: 4,
-    }), esAlias),
-  ]);
+  t.context.collectionPgModel = new CollectionPgModel();
+  t.context.granulePgModel = new GranulePgModel();
 
-  stub.restore();
+  const statuses = ['queued', 'failed', 'completed', 'running'];
+  const errors = [{ Error: { keyword: 'UnknownError' } }, { Error: { keyword: 'CumulusMessageAdapterError' } }, { Error: { keyword: 'IngestFailure' } }, { Error: { keyword: 'CmrFailure' } }, { Error: {} }];
+  const granules = [];
+  const collections = [];
+
+  range(20).map((num) => (
+    // collections is never aggregate queried
+    collections.push(fakeCollectionRecordFactory({
+      name: `testCollection${num}`,
+      cumulus_id: num,
+    }))
+  ));
+
+  range(100).map((num) => (
+    // granules can be queried by timestampto/from, collectionid, providerid, status,
+    granules.push(fakeGranuleRecordFactory({
+      collection_cumulus_id: num % 20,
+      status: statuses[num % 4],
+      created_at: (new Date(2018 + (num % 6), (num % 12), (num % 30))).toISOString(),
+      updated_at: (new Date(2018 + (num % 6), (num % 12), ((num + 1) % 29))).toISOString(),
+      error: errors[num % 5],
+      duration: num + (num / 10),
+    }))
+  ));
+
+  await t.context.collectionPgModel.insert(
+    t.context.knex,
+    collections
+  );
+
+  await t.context.granulePgModel.insert(
+    t.context.knex,
+    granules
+  );
 });
 
-test.after.always(async () => {
+test.after.always(async (t) => {
   await Promise.all([
-    esClient.client.indices.delete({ index: esIndex }),
     await accessTokenModel.deleteTable(),
     s3.recursivelyDeleteS3Bucket(process.env.system_bucket),
   ]);
+
+  await destroyLocalTestDb({
+    ...t.context,
+    testDbName,
+  });
 });
 
 test('GET without pathParameters and without an Authorization header returns an Authorization Missing response', async (t) => {
@@ -140,54 +160,6 @@ test('GET /stats/aggregate with an invalid access token returns an unauthorized 
   assertions.isInvalidAccessTokenResponse(t, response);
 });
 
-test('getType gets correct type for granules', (t) => {
-  const type = getType({ params: { type: 'granules' } });
-
-  t.is(type, 'granule');
-});
-
-test('getType gets correct type for collections', (t) => {
-  const type = getType({ params: { type: 'collections' } });
-
-  t.is(type, 'collection');
-});
-
-test('getType gets correct type for pdrs', (t) => {
-  const type = getType({ params: { type: 'pdrs' } });
-
-  t.is(type, 'pdr');
-});
-
-test('getType gets correct type for executions', (t) => {
-  const type = getType({ params: { type: 'executions' } });
-
-  t.is(type, 'execution');
-});
-
-test('getType gets correct type for logs', (t) => {
-  const type = getType({ params: { type: 'logs' } });
-
-  t.is(type, 'logs');
-});
-
-test('getType gets correct type for providers', (t) => {
-  const type = getType({ params: { type: 'providers' } });
-
-  t.is(type, 'provider');
-});
-
-test('getType returns undefined if type is not supported', (t) => {
-  const type = getType({ params: { type: 'provide' } });
-
-  t.falsy(type);
-});
-
-test('getType returns correct type from query params', (t) => {
-  const type = getType({ query: { type: 'providers' } });
-
-  t.is(type, 'provider');
-});
-
 test('GET /stats returns correct response, defaulted to all', async (t) => {
   const response = await request(app)
     .get('/stats')
@@ -195,23 +167,23 @@ test('GET /stats returns correct response, defaulted to all', async (t) => {
     .set('Authorization', `Bearer ${jwtAuthToken}`)
     .expect(200);
 
-  t.is(response.body.errors.value, 2);
-  t.is(response.body.collections.value, 2);
-  t.is(response.body.processingTime.value, 2.2);
-  t.is(response.body.granules.value, 5);
+  t.is(response.body.errors.value, '80');
+  t.is(response.body.processingTime.value, '54.44999999642372');
+  t.is(response.body.granules.value, '100');
+  t.is(response.body.collections.value, '20');
 });
 
 test('GET /stats returns correct response with date params filters values correctly', async (t) => {
   const response = await request(app)
-    .get(`/stats?timestamp__from=${(new Date(2020, 0, 28)).getTime()}&timestamp__to=${(new Date(2020, 0, 30)).getTime()}`)
+    .get(`/stats?timestamp__from=${(new Date(2018, 1, 28)).getTime()}&timestamp__to=${(new Date(2019, 1, 30)).getTime()}`)
     .set('Accept', 'application/json')
     .set('Authorization', `Bearer ${jwtAuthToken}`)
     .expect(200);
 
-  t.is(response.body.errors.value, 1);
-  t.is(response.body.collections.value, 1);
-  t.is(response.body.processingTime.value, 4);
-  t.is(response.body.granules.value, 2);
+  t.is(response.body.errors.value, '15');
+  t.is(response.body.collections.value, '10');
+  t.is(response.body.processingTime.value, '53.38235317258274');
+  t.is(response.body.granules.value, '17');
 });
 
 test('GET /stats/aggregate returns correct response', async (t) => {
@@ -221,21 +193,35 @@ test('GET /stats/aggregate returns correct response', async (t) => {
     .set('Authorization', `Bearer ${jwtAuthToken}`)
     .expect(200);
 
-  t.is(response.body.meta.count, 5);
-  t.deepEqual(response.body.count, [
-    { key: 'completed', count: 3 }, { key: 'failed', count: 2 },
-  ]);
+  const expectedResponse = {
+    meta: { name: 'cumulus-api', count: '100', field: 'status' },
+    count: [
+      { key: 'completed', count: '25' },
+      { key: 'running', count: '25' },
+      { key: 'queued', count: '25' },
+      { key: 'failed', count: '25' },
+    ],
+  };
+
+  t.deepEqual(response.body, expectedResponse);
 });
 
 test('GET /stats/aggregate filters correctly by date', async (t) => {
   const response = await request(app)
-    .get(`/stats/aggregate?type=granules&timestamp__from=${(new Date(2020, 0, 28)).getTime()}&timestamp__to=${(new Date(2020, 0, 30)).getTime()}`)
+    .get(`/stats/aggregate?type=granules&timestamp__from=${(new Date(2020, 11, 28)).getTime()}&timestamp__to=${(new Date(2023, 8, 30)).getTime()}`)
     .set('Accept', 'application/json')
     .set('Authorization', `Bearer ${jwtAuthToken}`)
     .expect(200);
 
-  t.is(response.body.meta.count, 2);
-  t.deepEqual(response.body.count, [
-    { key: 'completed', count: 1 }, { key: 'failed', count: 1 },
-  ]);
+  const expectedResponse = {
+    meta: { name: 'cumulus-api', count: '41', field: 'status' },
+    count: [
+      { key: 'failed', count: '16' },
+      { key: 'running', count: '9' },
+      { key: 'completed', count: '8' },
+      { key: 'queued', count: '8' },
+    ],
+  };
+
+  t.deepEqual(response.body, expectedResponse);
 });